backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
bcrypt==4.3.0
python-dotenv==1.0.1
gunicorn==23.0.0
